Programmatically creating ASFTableView
- Fixed programmatically creation of ASFTableView when used with Auto
Layouts
- Demo Added which creates another ASFTableView programmatically

+++ ASFTableViewDemo/ASFTableView/ASFLabel.m
@@ -9,7 +9,7 @@
 #import "ASFLabel.h"
 
 @interface ASFLabel () {
-    int padding;
+    NSUInteger padding;
 }
 
 @end
